+++ CHANGELOG.md
@@ -16,7 +16,6 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 - 🔴 Removed model `AutoARIMA`. To support `numpy>=2.0.0`, we unfortunately had to remove the `pmdarima` dependency. Use `StatsForecastAutoARIMA` instead. [#2734](https://github.com/unit8co/darts/pull/2734) by [Dennis Bader](https://github.com/dennisbader).
 - 🔴 Removed deprecated method `TimeSeries.pd_dataframe()`. Use `TimeSeries.to_dataframe()` instead. [#2733](https://github.com/unit8co/darts/pull/2733) by [Dennis Bader](https://github.com/dennisbader).
 - 🔴 Removed deprecated method `TimeSeries.pd_serise()`. Use `TimeSeries.to_series()` instead. [#2733](https://github.com/unit8co/darts/pull/2733) by [Dennis Bader](https://github.com/dennisbader).
-- 🔴 Likelihoods for `TorchForecastingModel` must now be imported from `darts.utils.likelihood_models.torch` instead of `darts.utils.likelihood_models`. [#2742](https://github.com/unit8co/darts/pull/2742) by [Dennis Bader](https://github.com/dennisbader).
 
 **Fixed**
 
@@ -26,9 +25,11 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 
 - Refactored likelihoods: [#2742](https://github.com/unit8co/darts/pull/2742) by [Dennis Bader](https://github.com/dennisbader).
   - Moved all likelihood related files into `darts/utils/likelihood_models/`.
-  - Added `BaseLikelihood` as a base class for all likelihood models.
-  - `RegressionModel` likelihoods are now handled by dedicated Likelihood models similar to `TorchForecastingModel`.
+  - Added `BaseLikelihood` as a base class for all likelihood models to `darts.utils.likelihood_models.likelihood`.
+  - Added `RegressionModel` likelihoods to `darts.utils.likelihood_models.sklearn`.
+  - Moved `TorchForecastingModel` likelihoods to `darts.utils.likelihood_models.torch`. They can still be imported through `darts.utils.likelihood_models`.
   - Removed `darts.models.forecasting.regression_model._LikelihoodMixin`. Use dedicated Likelihood models instead.
+
 
 ## [0.34.0](https://github.com/unit8co/darts/tree/0.34.0) (2025-03-09)